separate simd_aarch64
