+++ src/lib.rs
@@ -1,7 +1,6 @@
 #![doc = include_str!("../README.md")]
 
 mod errors;
-mod int_parser;
 mod jiter;
 mod lazy_index_map;
 mod number_decoder;
@@ -10,6 +9,8 @@ mod parse;
 mod py_string_cache;
 #[cfg(feature = "python")]
 mod python;
+#[cfg(target_arch = "aarch64")]
+mod simd_aarch64;
 mod string_decoder;
 mod value;
